fix success page in wrong language
If the user fills out the form in a language other than the deault,
the success page is still shown in the default language. This fix
selects the appropriate translation of the success page (and others
like the abroad redirect or the local initiative error redirect).
It is at least compatible with WPML and Polylang.